+++ systeminfo/multinode.go
@@ -60,10 +60,10 @@ func (m *MultinodeSystemInfoProvider) getComponents(ctx context.Context) ([]comp
 			// skip ces-exporter, otherwise the target instance tries to validate that the ces-exporter is installed
 			continue
 		}
-		slog.Debug(fmt.Sprintf("found component %s in version %s installed", c.Spec.Name, c.Spec.Version))
+		slog.Debug(fmt.Sprintf("found component %s in version %s installed", c.Spec.Name, c.Status.InstalledVersion))
 		components = append(components, component{
 			Name:    c.Spec.Name,
-			Version: c.Spec.Version,
+			Version: c.Status.InstalledVersion,
 		})
 	}